refactor: bookmark-table 컴포넌트와 스타일 분리 (#154)
